base: Fix narrowing casting for Musl

+++ src/base/files/file_util_linux.cc
@@ -23,7 +23,7 @@ bool GetFileSystemType(const FilePath& path, FileSystemType* type) {
 
   // Not all possible |statfs_buf.f_type| values are in linux/magic.h.
   // Missing values are copied from the statfs man page.
-  switch (statfs_buf.f_type) {
+  switch (static_cast<int>(statfs_buf.f_type)) {
     case 0:
       *type = FILE_SYSTEM_0;
       break;
